requirements.txt
aiogram>=3.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
httpx>=0.24.0
pydantic>=2.0.0
flask>=2.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
